requirements.txt
streamlit
reportlab
pyswisseph
requests
